Add Guerrilla struct and global New/Run interface for starting and running apps
